cast to uint

+++ commandhandler.c
@@ -144,7 +144,7 @@ static command_status_t command_handler(client_connection_t *client, char *cmd)
 	else if(!strncmp(cmd, "STATUS", 6))
 	{
 		char out[64];
-		int l = sprintf(out, "{con:%d,mp:%d,ps:%d}\n", server->connection_count, server->total_pixels_received/1000000, server->pixels_received_per_second);
+		int l = sprintf(out, "{con:%d,mp:%d,ps:%d}\n", server->connection_count, (uint) server->total_pixels_received/1000000, server->pixels_received_per_second);
 		send(client->socket, out, l, MSG_DONTWAIT | MSG_NOSIGNAL);
 		return COMMAND_SUCCESS;
 	}
